+++ CHANGES.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 0.22.0
 
 - Add rebellion soulbound reward cards. Portfolio and filter options
+- Add calculation of chaos legion soulbound for portfolio
 - FIX/CHANGE: remove pycoingecko dependency use hive price of SPL
 - FIX: Collection page is update only with current collection for determining portfolio value previous owned rented card are removed.